requirements.txt
flask==3.1.0
flask-cors==5.0.1
pandas==2.2.3
numpy==2.2.3
smartapi-python==1.5.5
pyotp==2.9.0
requests==2.32.3
logzero==1.7.0
python-dotenv==1.0.1